requirements.txt
moviepy==1.0.3
openai_whisper==20231117
selenium==4.16.0
